[nat64] NAT64 translator should be active only when local prefix is published
