+++ tests/scripts/thread-cert/border_router/nat64/test_multi_border_routers.py
@@ -26,7 +26,6 @@
 #  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 #  POSSIBILITY OF SUCH DAMAGE.
 #
-from typing import Mapping
 import unittest
 
 import config
@@ -85,19 +84,6 @@ class Nat64MultiBorderRouter(thread_cert.TestCase):
         },
     }
 
-    def assertDictIncludes(self, actual: Mapping[str, str], expected: Mapping[str, str]):
-        """ Asserts the `actual` dict includes the `expected` dict.
-
-        Args:
-            actual: A dict for checking.
-            expected: The expected items that the actual dict should contains.
-        """
-        for k, v in expected.items():
-            if k not in actual:
-                raise AssertionError(f"key {k} is not found in first dict")
-            if v != actual[k]:
-                raise AssertionError(f"{repr(actual[k])} != {repr(v)} for key {k}")
-
     def test(self):
         br1 = self.nodes[BR1]
         router = self.nodes[ROUTER]
@@ -123,6 +109,7 @@ class Nat64MultiBorderRouter(thread_cert.TestCase):
         #
         # Case 1. BR2 with an infrastructure prefix joins the network later and
         #         it will add the infrastructure nat64 prefix to Network Data.
+        #         Note: NAT64 translator will be bypassed.
         #
         br2.start()
         # When feature flag is enabled, NAT64 might be disabled by default. So
@@ -146,7 +133,7 @@ class Nat64MultiBorderRouter(thread_cert.TestCase):
         })
         self.assertDictIncludes(br2.nat64_state, {
             'PrefixManager': NAT64_STATE_ACTIVE,
-            'Translator': NAT64_STATE_ACTIVE
+            'Translator': NAT64_STATE_NOT_RUNNING
         })
 
         #
